core: replace fixed SMS body slots with variable-length pool
The previous design embedded body[SMS_BODY_MAX] directly inside each
SmsEntry, making M17_SMS_POOL_BYTES nothing more than a confusing way
to express a fixed per-slot limit of 64 bytes.  No variable-length
benefit was realised.

Replace with a real ring-buffer body pool:
- body_pool[M17_SMS_POOL_BYTES] holds NUL-terminated body strings
  back-to-back with a pool_head write cursor that wraps at POOL_BYTES.
- SmsEntry now stores (pool_offset, pool_len) instead of a body array.
- pool_alloc() advances pool_head and evicts any active entry whose
  stored region overlaps the new write range (those entries are always
  older because the cursor only moves forward).
- SmsEntry header slots (M17_SMS_MAX_MESSAGES = 64) remain a hard cap
  and continue to size CONFIG_MSG_SNAPSHOT_SIZE.

This means short messages pack tightly; a 100-char average fills the
6400-byte pool with ~63 messages before eviction begins.  A rare
821-char message simply uses more pool space without wasting 821 bytes
across every slot.

Also fix two related bottlenecks uncovered by the pool work:
- M17_SMS_POOL_BYTES raised from 4096 to 6400 (64 x 100-char average)
  on all targets; missing macros added to DM-1701 and Module17.
- ui_state_t::input_position widened from uint8_t to uint16_t, and
  _ui_textInputKeypad max_len from uint8_t to uint16_t, so the compose
  body editor can accept up to 821 characters (protocol maximum).
  The MESSAGES_COMPOSE_BODY keypad call updated from 254 to 821.
- rx_text parse buffer made function-scope static to avoid 822-byte
  stack allocation on the embedded RTX thread.

+++ openrtx/include/ui/ui_default.h
@@ -214,7 +214,7 @@ typedef struct ui_state_t
     bool input_locked;
     // Variables used for VFO input
     uint8_t input_number;
-    uint8_t input_position;
+    uint16_t input_position;
     uint8_t input_set;
     long long last_keypress;
     freq_t new_rx_frequency;
